Add a check and clear function * In this branch i have added a function to mark a item complete, * have added a function to allow user to clear the item from the list, * the user also has the ability to clear all the completed items.
