+++ src/index.js
@@ -1,5 +1,6 @@
 import './index.css';
 import './components/add-delete.js';
+import './components/clear.js';
 import '@fortawesome/fontawesome-free/js/fontawesome.js';
 import '@fortawesome/fontawesome-free/js/solid.js';
 import '@fortawesome/fontawesome-free/js/regular.js';
